Added team entity to group users under a manager in the system, leave request workflows updated accordingly need more crosschecking

+++ src/main/java/com/elms/elms_backend/dto/leave_balance/EmployeeLeaveBalanceSummaryDTO.java
@@ -8,9 +8,10 @@ import lombok.NoArgsConstructor;
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
+
 public class EmployeeLeaveBalanceSummaryDTO {
 private Long id;
 private String name;
-private Integer totalAllocated;
-private Integer totalRemaining;
+private Long totalAllocated;
+private Long totalRemaining;
 }
